refactor(dialogs): Move dialog components in dialog component folder, renaming, and import fixes
