+++ src/components/dialogs/create-contingency-list-dialog.js
@@ -19,13 +19,13 @@ import Alert from '@material-ui/lab/Alert';
 
 import { FormattedMessage, useIntl } from 'react-intl';
 
-import { createContingencyList, elementExists } from '../utils/rest-api';
+import { createContingencyList, elementExists } from '../../utils/rest-api';
 
 import { useSelector } from 'react-redux';
 import RadioGroup from '@material-ui/core/RadioGroup';
 import Radio from '@material-ui/core/Radio';
 import PropTypes from 'prop-types';
-import { ContingencyListType, ElementType } from '../utils/elementType';
+import { ContingencyListType, ElementType } from '../../utils/elementType';
 import CircularProgress from '@material-ui/core/CircularProgress';
 import { makeStyles } from '@material-ui/core/styles';
 import CheckIcon from '@material-ui/icons/Check';
@@ -37,7 +37,7 @@ const useStyles = makeStyles(() => ({}));
  * @param {Boolean} open Is the dialog open ?
  * @param {EventListener} onClose Event to close the dialog
  */
-export const CreateContingencyListForm = ({ open, onClose }) => {
+export const CreateContingencyListDialog = ({ open, onClose }) => {
     const [contingencyListType, setContingencyListType] = React.useState(
         ContingencyListType.SCRIPT
     );
@@ -250,7 +250,9 @@ export const CreateContingencyListForm = ({ open, onClose }) => {
                             }
                             type="text"
                             style={{ width: '90%' }}
-                            label=<FormattedMessage id="contingencyListName" />
+                            label={
+                                <FormattedMessage id="contingencyListName" />
+                            }
                         />
                         {renderContingencyNameStatus()}
                     </div>
@@ -262,7 +264,9 @@ export const CreateContingencyListForm = ({ open, onClose }) => {
                         value={contingencyListDescription}
                         type="text"
                         style={{ width: '90%' }}
-                        label=<FormattedMessage id="contingencyListDescription" />
+                        label={
+                            <FormattedMessage id="contingencyListDescription" />
+                        }
                     />
 
                     <RadioGroup
@@ -295,12 +299,12 @@ export const CreateContingencyListForm = ({ open, onClose }) => {
                         <FormControlLabel
                             value="public"
                             control={<Radio />}
-                            label=<FormattedMessage id="public" />
+                            label={<FormattedMessage id="public" />}
                         />
                         <FormControlLabel
                             value="private"
                             control={<Radio />}
-                            label=<FormattedMessage id="private" />
+                            label={<FormattedMessage id="private" />}
                         />
                     </RadioGroup>
                     {createContingencyListErr !== '' && (
@@ -330,9 +334,9 @@ export const CreateContingencyListForm = ({ open, onClose }) => {
     );
 };
 
-CreateContingencyListForm.propTypes = {
+CreateContingencyListDialog.propTypes = {
     open: PropTypes.bool.isRequired,
     onClose: PropTypes.func.isRequired,
 };
 
-export default CreateContingencyListForm;
+export default CreateContingencyListDialog;
